requirements.txt
openai
ollama
streamlit
python-dotenv
pandas